Changes needed for normalizing data shapes
I made significant changes to how the data is parsed so that data shapes
can be normalized to the largest dimensions. This means that all data
will be grouped together for fitting as opposed to fitting groups of
differing record sizes.

The main concern here was differing size lag tables, and missing range
gates(slist<numranges)

To account for differing lag tables, an array is allocated for the
largest lag table. Any smaller lag table will be inserted in its
location and the unused indices are masked. The maximum size lag
alternate lag 0 will be trimmed, but alt lag 0s are masked here so that
the alt lag 0 of smaller tables isnt used in processing

Using slist the data mask is also updated to mark where range data is
missing.