Changes based of real data
When testing with real data a couple issues needed solving

Real data showed that its possible for the gradients to move to a point
where the matrix becomes noninvertable without converging. The code was
changed to detect this and mark a fit as bad if this condition happens.

The other issue is that rho could be zero without converging. This means
the algo is calculating the same chi2 so this condition is also checked.